+++ tlwatch/cli.py
@@ -2,6 +2,7 @@ import click
 import tlwatch.etherscan
 import tlwatch.jsonrpc
 import tlwatch.relay
+import tlwatch.psql
 
 
 @click.group()
@@ -12,3 +13,4 @@ def cli():
 cli.add_command(tlwatch.etherscan.etherscan)
 cli.add_command(tlwatch.jsonrpc.jsonrpc)
 cli.add_command(tlwatch.relay.relay)
+cli.add_command(tlwatch.psql.psql)